Big cleanup + namespacing + addition of lazy_load.
