+++ .pre-commit-config.yaml
@@ -8,7 +8,7 @@ repos:
       - id: check-yaml
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.7.2
+    rev: v0.11.5
     hooks:
       - id: ruff
         args: [--fix]
