+++ tox.toml
@@ -1,0 +1,20 @@
+requires = ["tox>=4.24.1"]
+env_list = ["coverage_clean", "3.9", "3.10", "3.11", "3.12", "3.13", "3.13t", "pypy3", "coverage"]
+skip_missing_interpreters = true
+
+[env_run_base]
+package = "wheel"
+wheel_build_env = ".pkg"
+dependency_groups = ["test", "cov"]
+commands = [["coverage", "run", "-m", "pytest", { replace = "posargs", extend = true }]]
+
+[env.coverage_clean]
+skip_install = true
+dependency_groups = ["cov"]
+commands = [["coverage", "erase"]]
+
+[env.coverage]
+skip_install = true
+dependency_groups = ["cov"]
+depends = ["3.9", "3.10", "3.11", "3.12", "3.13", "3.13t", "pypy3"]
+commands = [["coverage", "combine"], ["coverage", "report"]]
